Move logging to log_config
