+++ scripts/tests/chiptest/log_config.py
